Gave Discord bot its own file, tweaked README
